+++ CHANGES.md
@@ -1,3 +1,8 @@
 - Add filter settings for max cost per knowledge and artisan currency to NPC orders page
 - Make "Track All" only track unprofitable orders that give enough knowledge or currency
-- Internal: Dramatically reduce attributed memory usage
+- Update concentration cost when adding extra skill in experimentation mode
+- Show correct concentration cost for optimized unclaimed orders
+- Hide optimization buttons in linked profession windows
+- Optimize orders in order list asynchronously, to prevent frame drops and errors
+- Fix not reoptimizing allocations when adding extra skill in experimentation mode
+- Fix error when optmizing recipe with bonus skill slot unlocked